Move some files to inner plans
