+++ tests/unit_tests/common/external_interaction/xray_centre/test_ispyb_callback.py
@@ -5,7 +5,9 @@ from bluesky.preprocessors import run_decorator, subs_decorator
 from ophyd_async.core import init_devices
 from ophyd_async.epics.core import epics_signal_rw
 
-from mx_bluesky.common.experiment_plans.read_hardware import read_hardware_plan
+from mx_bluesky.common.experiment_plans.inner_plans.read_hardware import (
+    read_hardware_plan,
+)
 from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback import (
     GridscanISPyBCallback,
 )
